move transactionData to ../transaction-data.js and update test to assert.rejects
