+++ test/contracts/transaction-data/TransactionData.sol
@@ -2,6 +2,6 @@ pragma solidity ^0.4.2;
 
 contract TransactionData {
     function () external payable {
-        require(msg.value > 0 && msg.data.length == 0);
+        require(msg.data.length == 0, "msg.data.length was 0");
     }
 }
